fix UnhandledPromiseRejection (#9)
this was quite a struggle. since we wanted to `await sleep()` a certain
amount of time after kicking of the requests, we received an
UnhandledPromiseRejection because the request-promises were rejected
while sleeping. This fix introduces some boilerplate code, but gets rid
of the issue in a straightforward manner.

See: https://stackoverflow.com/questions/59060508/how-to-handle-an-unhandled-promise-rejection-asynchronously
Also related: https://stackify.com/node-js-error-handling/
Also related: https://www.joyent.com/node-js/production/design/errors.